requirements.txt
Django==3.0.3
djangorestframework==3.11.0
django-dotenv==1.4.2
